requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.11.12
    # via requests
cffi==2.0.0
    # via pynacl
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via flask
flask==3.1.2
    # via -r requirements.in
idna==3.11
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
pycparser==2.23
    # via cffi
pynacl==1.6.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.5.0
    # via requests
werkzeug==3.1.4
    # via flask
